requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.5.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.0
flatbuffers==2.0.7
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.0.3
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-multipart==0.0.9
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
